requirements.txt
flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
paho-mqtt==1.6.1
python-dateutil==2.8.2
pytz==2021.1
werkzeug==2.0.3 